COM-1151 change seed
